Fixes serious bugs in reading A2M format alignments.
esl_msafile_a2m_Read() was misparsing any A2M line ending in an
insertion, because of an off-by-one error that overwrote the insertion
length counter this_nins[]. Fixed.

Unacceptable that this escaped testing; the case is not exercised by
any unit test.

Added utest_gibberish() to esl_msafile_a2m, to test read/write of
sampled dirty MSAs. Added esl_msa_Sample() to stochastically generate
pathologically dirty MSAs; and utest_Sample() to test it.

Added esl_msa_FlushLeftInserts(), to allow unit test
comparisons to succeed, because an input A2M file will be read with
all its inserts flushed left.

utest_gibberish() detects additional bugs in parsing all-insertion and
zero-length sequences; fixed both.

Bug was reported by Mohammed AlQuraishi [hms.harvard.edu], 19 Apr
2017.

+++ esl_alphabet.h
@@ -118,10 +118,10 @@ extern int    esl_abc_ValidateSeq(const ESL_ALPHABET *a, const char *seq, int64_
 #define esl_abc_CIsUnknown(a, c)     ((a)->inmap[(int)c] == (a)->Kp-3)
 #define esl_abc_CIsNonresidue(a, c)  ((a)->inmap[(int)c] == (a)->Kp-2)
 #define esl_abc_CIsMissing(a, c)     ((a)->inmap[(int)c] == (a)->Kp-1)
-#define esl_abc_CGetGap(a)           ((a)->sym[(int)(a)->K])
-#define esl_abc_CGetUnknown(a)       ((a)->sym[(int)(a)->Kp-3])
-#define esl_abc_CGetNonresidue(a)    ((a)->sym[(int)(a)->Kp-2])
-#define esl_abc_CGetMissing(a)       ((a)->sym[(int)(a)->Kp-1])
+#define esl_abc_CGetGap(a)           ((a)->sym[(a)->K])
+#define esl_abc_CGetUnknown(a)       ((a)->sym[(a)->Kp-3])
+#define esl_abc_CGetNonresidue(a)    ((a)->sym[(a)->Kp-2])
+#define esl_abc_CGetMissing(a)       ((a)->sym[(a)->Kp-1])
 
 #endif /*eslALPHABET_INCLUDED*/
 /*****************************************************************
